Move PPOPolicy to OnlineActorCriticPolicy
