+++ stable_baselines3/a2c/__init__.py
@@ -1,2 +1,2 @@
 from stable_baselines3.a2c.a2c import A2C
-from stable_baselines3.ppo.policies import MlpPolicy, CnnPolicy
+from stable_baselines3.common.policies import OnlineActorCriticPolicy, OnlineActorCriticCnnPolicy
