+++ stable_baselines3/ppo/__init__.py
@@ -1,2 +1,2 @@
 from stable_baselines3.ppo.ppo import PPO
-from stable_baselines3.ppo.policies import MlpPolicy, CnnPolicy
+from stable_baselines3.common.policies import OnlineActorCriticPolicy, OnlineActorCriticCnnPolicy
